mbs_make now works without warnings on macosx
git-svn-id: https://svn.nfit.au.dk/svn/beta/trunk@13198 678b6752-a8a0-49ff-a511-7c2a3241fcd4

+++ betarun/NEWRUN/AllocateComponent.c
@@ -17,7 +17,7 @@ Component *AlloC(Object *origin, ProtoType *proto, long *SP)
   
   DEBUG_ALLOI(fprintf(output,
 		      "AlloC#%d: origin=0x%x (%s), proto=0x%x (%s)\n",
-		      NumAlloC,
+		      (int)NumAlloC,
 		      (int)origin, 
 		      origin ? ProtoTypeName(GETPROTO(origin)) : 0, 
 		      (int)proto,
